fix(Menu): fix broken alignments link in trach hover menu
The link was previously pointed to a page that does not exist on EG. With this change it is pointing to http://ensemblgenomes.org/info/data/whole_genome_alignment.

Also add a link to this page http://plants.ensembl.org/mlss.html?mlss=<mlss_id> onto the same menu.

Related to ENSEMBL-4490.