feat(Inline::C): add a new script to compile the Inline::C code
We don't have a C compiler in our live environment due to Web Production security policy so we have to make sure the inline C code is compiled in the staging environment and copied to the live nodes.

This new script triggers all the Inline::C code to be built for the current site.

Related to ENSEMBL-4496.