Use the old species(/alignment) selector in alignments view
From Ensembl release/86 there is a new interface to select pairwise alignments.
We need to use the old one before figuring out how this selector should look
like for Genomes.

Copy related module from Ensembl release/85 to eg-web-common so that the old one
is still in use.

Related to ENSEMBL-4696.