Add new species selector
Currently it is tested with the species selector in Blast input form at Plants site.
New json file htdocs/e_species_divisions.json should be copied to ensembl-webcode/htdocs
to make it work. Also this new file should live under eg-web-plants/htdocs. Leave it
here just for easier to manage this feature branch. The new json file is currently
manually created. In the future this should be created through a script that is similar
to the one to generate taxon tree.

Related to ENSEMBL-4808.