Add copy-on-write feature when processing species ini file
By default, when processing species ini files, if a new section has been defined in
the defaults.ini, it will create a deep copy of that from the defaults.ini. This is
not space efficient for EG as there are two big hash tables [ENSEMBL_SPECIES_SITE]
and [SPECIES_DISPLAY_NAME]. Having these two tables for each species/collection will
waste a lot of space if there is no overwrite from the species ini.

Now the copy-on-write is only enabled for those two and [ENSEMBL_EXTERNAL_URLS] as
it seems difficult to make it work for all other sections. There will be problems
with [databases] because of the way databases ini are handled currently.