Fix copy-on-write behaviour for species defs
In commit f702ab7 we introduce copy-on-write for three sections when parsing
species ini files. But there is a problem after these ini files are parsed.
The parsed tree is merged with db_tree stored in the .db.packed file and the
ones in the db_tree take precedence and hence overwriting the cow behaviour
of those three sections for those species that have related .db.packed files.
Hence for those species, there would be a deep copy for those three sections.

This change will fix this issue by preferring the ones from parsed tree for
those three sections. It is a quick hack, future refactoring is desired.