Remove child_terminate call from http_redirect
With current implementation of ensembl-webcode, child_terminate is called when doing
http_redirect. This hurts server performance badly because the EG server processes
tend to be heavy (require large momory), starting and stoping a worker process is
an expensive operation.

We overwrite the behaviour from ensembl-webcode here so that EG servers can benefit
from this.