fix: check if a key is defined before dereferencing it
Due to Perl's autovivification feature, if a key is not defined, dereferencing
it will automaticaly create an empty hash as value for that key. This will be
an issue when checking if the key is defined elsewhere.

This change remove this potential problem for 'DATABASE_FUNCGEN' and 'DATABASE
_VARIATION' keys in species_defs->databases.

Related to ENSEMBL-4611.